feat(map/basic): removed search-map middleware
Delegated all code bunch to leaflet-map class
